+++ doc/whats_new.rst
@@ -22,6 +22,7 @@ The following authors contributed for the first time. Thank you so much! 🤩
 * `Julius Welzel`_
 * `Alex Lopez Marquez`_
 * `Bruno Aristimunha`_
+* `Kalle Mäkelä`_
 
 The following authors had contributed before. Thank you for sticking around! 🤘
 
@@ -44,6 +45,7 @@ Detailed list of changes
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 -  `tracksys` accepted as argument in :class:`mne_bids.BIDSPath()` by `Julius Welzel`_ (:gh:`1430`)
+- :func:`mne_bids.read_raw_bids()` has a new parameter ``on_ch_mismatch`` that controls behaviour when there is a mismatch between channel names in ``channels.tsv`` and the raw data; accepted values are ``'raise'`` (default), ``'reorder'``, and ``'rename'``, by `Kalle Mäkelä`_.
 
 🛠 Requirements
 ^^^^^^^^^^^^^^^